Removed unstable functions from ccd_tools, in preperation for deployment Placed these functions in ds9_region_methods.py
Wrote an extensive README, which is intended to get anyone up and
running with this module. Also added a docstring to the module, listing
all the functions and classes.
